backend/requirements.txt
fastapi
uvicorn[standard]
email-validator
python-dotenv
pyjwt
supabase
